feat: introduce `max_threads` configuration
This commit introduces a new `max_threads` configuration option to the application.
This option allows users to specify the maximum number of threads that the application will use for parallel processing.
By default, the application will use the number of available CPU cores, but users can override this value if desired.