feat: Allow throttle and debounce to be used concurrently
Previously, the throttle logic would only be applied if debounce was set to 0. This change removes that condition, allowing the throttle mechanism to operate even when debounce is configured.

This enables scenarios where both debounce (to limit rapid triggers) and throttle (to limit execution frequency after a trigger) are desired.